refactor(forum): migration BlockedPost à forum_moderation
