+++ lacommunaute/forum_conversation/forms.py
@@ -5,7 +5,8 @@ from machina.apps.forum_conversation.forms import PostForm as AbstractPostForm, 
 from machina.conf import settings as machina_settings
 from taggit.models import Tag
 
-from lacommunaute.forum_conversation.models import BlockedPost, Post
+from lacommunaute.forum_conversation.models import Post
+from lacommunaute.forum_moderation.models import BlockedPost
 from lacommunaute.forum_moderation.utils import BlockedPostReason, check_post_approbation
