+++ lacommunaute/forum_moderation/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
-from lacommunaute.forum_moderation.models import BlockedDomainName, BlockedEmail
+from lacommunaute.forum_moderation.models import BlockedDomainName, BlockedEmail, BlockedPost
 
 
 @admin.register(BlockedEmail)
@@ -13,3 +13,9 @@ class BlockedEmailAdmin(admin.ModelAdmin):
 class BlockedDomainNameAdmin(admin.ModelAdmin):
     list_display = ("domain", "created", "reason")
     list_filter = ("reason",)
+
+
+@admin.register(BlockedPost)
+class BlockedPostAdmin(admin.ModelAdmin):
+    list_display = ("username", "created", "block_reason")
+    list_filter = ("block_reason",)
